requirements.txt
absl-py==2.1.0
acme==2.10.0
apiclient==1.0.4
array_record==0.5.1
astunparse==1.6.3
attrs==21.2.0
Authlib==1.3.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
bidict==0.23.1
blinker==1.7.0
cachelib==0.13.0
cachetools==5.3.3
certbot==2.10.0
certbot-apache==2.10.0
certbot-nginx==1.21.0
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
click==8.1.7
cloud-init==23.4.4
colorama==0.4.4
command-not-found==0.3
config==0.5.1
ConfigArgParse==1.5.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.2.1
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
dm-tree==0.1.8
ec2-hibinit-agent==1.0.0
etils==1.7.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Dance==7.1.0
Flask-Mail==0.9.1
Flask-MySQLdb==2.0.0
Flask-OAuthlib==0.9.6
Flask-SocketIO==5.3.6
Flask-WTF==1.2.1
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
gast==0.5.4
gevent==24.2.1
gevent-websocket==0.10.1
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.64.0
h11==0.14.0
h5py==3.11.0
hibagent==1.0.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
importlib_resources==6.4.0
incremental==21.3.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.3
josepy==1.14.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
ml-dtypes==0.3.2
mod-wsgi==5.0.0
more-itertools==8.10.0
mysqlclient==2.2.4
namex==0.0.8
netifaces==0.11.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==2.1.0
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
parsedatetime==2.6
pexpect==4.8.0
pillow==10.3.0
promise==2.3
proto-plus==1.23.0
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycparser==2.22
Pygments==2.18.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyICU==2.8.1
PyJWT==2.3.0
PyMySQL==1.1.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu3
python-augeas==1.1.0
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-engineio==4.9.1
python-magic==0.4.24
python-socketio==5.11.2
pytz==2022.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.1.0
requests-toolbelt==0.9.1
rfc3339==6.2
rich==13.7.1
rsa==4.9
SecretStorage==3.3.1
service-identity==18.1.0
simple-websocket==1.0.0
six==1.16.0
sos==4.5.6
ssh-import-id==5.11
systemd-python==234
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-datasets==4.9.4
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-metadata==1.15.0
termcolor==2.4.0
tf-estimator-nightly==2.8.0.dev2021122109
toml==0.10.2
tqdm==4.66.4
Twisted==22.1.0
typing_extensions==4.11.0
tzdata==2024.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
URLObject==2.4.3
wadllib==1.3.6
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.1.0
zope.interface==5.4.0
